more code cleanup

+++ src/cs2/triggerbot.cpp
@@ -51,7 +51,7 @@ void Triggerbot() {
     const f32 mean =
         static_cast<f32>(config.triggerbot.delay_min + config.triggerbot.delay_max) / 2.0f;
     std::normal_distribution<f32> normal{
-        mean, (config.triggerbot.delay_max - config.triggerbot.delay_min) / 2.0f};
+        mean, static_cast<f32>(config.triggerbot.delay_max - config.triggerbot.delay_min) / 2.0f};
 
     const i32 delay = static_cast<i32>(normal(rng));
     next_shot = std::chrono::steady_clock::now() + std::chrono::milliseconds(delay);
